+++ R/block.R
@@ -331,10 +331,7 @@ call_inline = function(block) {
   in_dir(opts_knit$get('root.dir') %n% input_dir(), inline_exec(block))
 }
 
-inline_exec = function(
-  block, eval = eval_lang(opts_chunk$get('eval')), envir = knit_global(),
-  hook = knit_hooks$get('inline')
-) {
+inline_exec = function(block, envir = knit_global(), hook = knit_hooks$get('inline')) {
 
   # run inline code and substitute original texts
   code = block$code; input = block$input
@@ -342,10 +339,8 @@ inline_exec = function(
 
   loc = block$location
   for (i in 1:n) {
-    res = if (eval) {
-      v = withVisible(eval(parse_only(code[i]), envir = envir))
-      if (v$visible) knit_print(v$value, inline = TRUE, options = opts_chunk$get())
-    } else '??'
+    v = withVisible(eval(parse_only(code[i]), envir = envir))
+    res = if (v$visible) knit_print(v$value, inline = TRUE, options = opts_chunk$get())
     if (inherits(res, c('knit_asis', 'knit_asis_list'))) res = wrap(res, inline = TRUE)
     d = nchar(input)
     # replace with evaluated results