+++ tests/testit/test-params.R
@@ -24,7 +24,7 @@ assert(params[[2]]$value == 20)
 
 assert(identical(flatten_params(params), list(a = 10L, b = 20L)))
 
-## test date custom type ---------------------------------------------------
+## test date custom type (these deprecated and here for backwards compt) --
 
 params <- read_params('
 ---
@@ -35,25 +35,11 @@ params:
 '
 )
 assert(params[[1]]$name == 'start')
-assert(params[[1]]$type == 'date')
+assert('Date' %in% params[[1]]$class)
 assert(params[[1]]$value == as.Date("2015-01-01"))
 assert(params[[2]]$name == 'end')
-assert(params[[2]]$type == 'datetime')
+assert('POSIXct' %in% params[[2]]$class)
 assert(params[[2]]$value == as.POSIXct("2015-01-01 12:30:00", tz = "GMT"))
-
-
-## test file custom type ---------------------------------------------------
-
-params <- read_params('
----
-params:
-  file: !file data.csv
----
-'
-)
-assert(params[[1]]$name == 'file')
-assert(params[[1]]$type == 'file')
-assert(params[[1]]$value == 'data.csv')
 
 
 ## test specifying value in sub-object and type at object level ------------
@@ -62,18 +48,12 @@ params <- read_params('
 ---
 params:
   file1:
-    value: !file data1.csv
-  file2: !file
-    value: data2.csv
+    value: data1.csv
 ---
 '
 )
 assert(params[[1]]$name == 'file1')
-assert(params[[1]]$type == 'file')
 assert(params[[1]]$value == 'data1.csv')
-assert(params[[2]]$name == 'file2')
-assert(params[[2]]$type == 'file')
-assert(params[[2]]$value == 'data2.csv')
 
 
 ## test parameters with length(value) > 1 ----------------------------------
@@ -125,3 +105,27 @@ assert(
   'y/Y/n/N are not converted to booleans',
   identical(unlist(lapply(params, `[[`, 'name')), c('x', 'y', 'z', 'n', 'Y', 'N'))
 )
+
+
+## test handling of expressions --------------------------------------------
+
+params <- read_params('
+---
+params:
+  today: !r Sys.Date()
+  now: !expr Sys.time()
+  x: 10
+---
+'
+)
+assert(!is.null(params[[1]]$expr))
+assert('Date' %in% params[[1]]$class)
+assert(params[[2]]$expr)
+assert('POSIXct' %in% params[[2]]$class)
+assert(is.null(params[[3]]$expr))
+
+
+
+
+
+