+++ R/engine.R
@@ -104,8 +104,8 @@ eng_interpreted = function(options) {
     )
   } else paste(switch(
     engine, bash = '-c', coffee = '-e', groovy = '-e', lein = 'exec -e',
-    node = '-e', perl = '-e', python = '-c', ruby = '-e', scala = '-e',
-    sh = '-c', zsh = '-c', mysql = '-e', psql = '-c', NULL
+    mysql = '-e', node = '-e', perl = '-e', psql = '-c', python = '-c',
+    ruby = '-e', scala = '-e', sh = '-c', zsh = '-c', NULL
   ), shQuote(paste(options$code, collapse = '\n')))
 
   # FIXME: for these engines, the correct order is options + code + file
@@ -291,8 +291,9 @@ eng_asis = function(options) {
 # set engines for interpreted languages
 local({
   for (i in c(
-    'awk', 'bash', 'coffee', 'gawk', 'groovy', 'haskell', 'lein', 'node', 'perl',
-    'python', 'Rscript', 'ruby', 'sas', 'scala', 'sed', 'sh', 'stata', 'zsh', 'mysql', 'psql'
+    'awk', 'bash', 'coffee', 'gawk', 'groovy', 'haskell', 'lein', 'mysql',
+    'node', 'perl', 'psql', 'python', 'Rscript', 'ruby', 'sas', 'scala', 'sed',
+    'sh', 'stata', 'zsh'
   )) knit_engines$set(setNames(list(eng_interpreted), i))
 })